requirements.txt
numpy
pyspark==2.4.6
pytorch==1.8.0
pytorch-lightning==1.4.9
pandas
torchvision==0.9.2
jupyterlab
Pillow
cudatoolkit==11.1
torchaudio==0.8.2
torchtext==0.9
huggingface